+++ src/renderer/vertex_layout.hpp
@@ -25,11 +25,11 @@ struct PositionNormalUvVertex {
 };
 
 struct PositionNormalUvBoneVertex {
-    glm::vec3  Position;
-    glm::vec3  Normal;
-    glm::vec2  Uv;
-    glm::vec4  BoneWeights{-1};
-    glm::ivec4 BoneIndices{-1};
+    glm::vec3 Position;
+    glm::vec3 Normal;
+    glm::vec2 Uv;
+    glm::vec4 BoneWeights{-1};
+    glm::vec4 BoneIndices{-1};
 
     static bgfx::VertexLayout GetVertexLayout()
     {
